Moves repository tag files out of dashboard folder
