+++ libs/client/dashboard/src/lib/components/tags-management/child-tag-item/child-tag-item.component.ts
@@ -1,5 +1,5 @@
 import { Component, Input } from '@angular/core';
-import { TagsService } from '@dragonfish/client/repository/dashboard/tags';
+import { TagsService } from '@dragonfish/client/repository/tags';
 import { AlertsService } from '@dragonfish/client/alerts';
 import { MatDialog } from '@angular/material/dialog';
 import { PopupModel } from '@dragonfish/shared/models/util';
